+++ src/lpj/drain.c
@@ -50,7 +50,9 @@ void drain(Cell grid[],         /**< Cell array */
   int count,cell,i,j,iter,t,ncoeff;
   Real fin,*out,*in;
   Real fout_lake,irrig_to_river;
-
+#ifdef USE_TIMING
+  double tstart;
+#endif
   count=(int)(1.0/TSTEP); /* calculate number of iterations */
   out=(Real *)pnet_output(config->route);
   in=(Real *)pnet_input(config->route);
@@ -115,6 +117,12 @@ void drain(Cell grid[],         /**< Cell array */
 
   grid-=config->startgrid-config->firstgrid; /* adjust first index of grid
                                               array needed for pnet library */
+#ifdef USE_TIMING
+#ifdef USE_MPI
+  MPI_Barrier(config->comm);
+#endif
+  tstart=mrun();
+#endif
   for(iter=0;iter<count;iter++)
   {
     for(i=pnet_lo(config->route);i<=pnet_hi(config->route);i++)
@@ -228,5 +236,7 @@ void drain(Cell grid[],         /**< Cell array */
     }
     getoutput(&grid[cell].output,WD_LOCAL,config)+=grid[cell].discharge.withdrawal/grid[cell].coord.area; /* withdrawal in local cell */
   }
-
+#ifdef USE_TIMING
+  timing.drain+=mrun()-tstart;
+#endif
 }  /* of 'drain' */