+++ src/lpjml.c
@@ -95,7 +95,11 @@ int main(int argc,char **argv)
   int year,rc;
   Cell *grid;         /* cell array */
   Input input;        /* input data */
-  time_t tstart,tend,tbegin,tfinal;   /* variables for timing */
+  time_t tstart,tend,tinvoke;   /* variables for timing */
+  double tbegin,tfinal;
+#ifdef USE_TIMING
+  double t;
+#endif
   Standtype standtype[NSTANDTYPES];
   String s;
   Config config;         /* LPJ configuration */
@@ -107,9 +111,13 @@ int main(int argc,char **argv)
     {name_tree,fscanpft_tree},
     {name_crop,fscanpft_crop}
   };
-  time(&tbegin);         /* Start timing for total wall clock time */
+  time(&tinvoke);
+  tbegin=mrun();         /* Start timing for total wall clock time */
 #ifdef USE_MPI
   MPI_Init(&argc,&argv); /* Initialize MPI */
+#ifdef USE_TIMING
+  timing.MPI_Init+=mrun()-tbegin;
+#endif
 /*
  * Use default communicator containing all processors. In defining your own
  * communicator it is possible to run LPJ on a subset of processors
@@ -174,7 +182,7 @@ int main(int argc,char **argv)
   { /* Output only for the root task 0 */
     copyright(progname);
     printf("\nRunning for user %s on %s at %s",getuser(),gethost(),
-           ctime(&tbegin));
+           ctime(&tinvoke));
     fflush(stdout);
   }
 
@@ -182,7 +190,13 @@ int main(int argc,char **argv)
    * in light and establishment
    * crops must have last id-number */
   /* Read configuration file */
+#ifdef USE_TIMING
+  t=mrun();
+#endif
   rc=readconfig(&config,scanfcn,NTYPES,NOUT,&argc,&argv,lpj_usage);
+#ifdef USE_TIMING
+  timing.readconfig+=mrun()-t;
+#endif
   failonerror(&config,rc,READ_CONFIG_ERR,"Cannot read configuration");
   if(argc)
   {
@@ -240,7 +254,13 @@ int main(int argc,char **argv)
   standtype[WOODPLANTATION]=woodplantation_stand;
   standtype[KILL]=kill_stand;
   /* Allocation and initialization of grid */
+#ifdef USE_TIMING
+  t=mrun();
+#endif
   rc=((grid=newgrid(&config,standtype,NSTANDTYPES,config.npft[GRASS]+config.npft[TREE],config.npft[CROP]))==NULL);
+#ifdef USE_TIMING
+  timing.newgrid+=mrun()-t;
+#endif
   failonerror(&config,rc,INIT_GRID_ERR,"Initialization of LPJ grid failed");
   if(iscoupled(config))
   {
@@ -248,7 +268,13 @@ int main(int argc,char **argv)
     snprintf(s,STRING_LEN,"Cannot couple to %s model",config.coupled_model);
     failonerror(&config,rc,OPEN_COUPLER_ERR,s);
   }
+#ifdef USE_TIMING
+  t=mrun();
+#endif
   rc=initinput(&input,grid,config.npft[GRASS]+config.npft[TREE],&config);
+#ifdef USE_TIMING
+  timing.initinput+=mrun()-t;
+#endif
   failonerror(&config,rc,INIT_INPUT_ERR,
               "Initialization of input data failed");
   if(config.check_climate)
@@ -257,11 +283,23 @@ int main(int argc,char **argv)
     failonerror(&config,rc,INIT_INPUT_ERR,"Check of climate data failed");
   }
   /* open output files */  
+#ifdef USE_TIMING
+  t=mrun();
+#endif
   output=fopenoutput(grid,NOUT,&config);
+#ifdef USE_TIMING
+  timing.fopenoutput+=mrun()-t;
+#endif
   rc=(output==NULL);
   failonerror(&config,rc,INIT_OUTPUT_ERR,
               "Initialization of output data failed");
+#ifdef USE_TIMING
+  t=mrun();
+#endif
   rc=initoutput(output,grid,config.npft[GRASS]+config.npft[TREE],config.npft[CROP],&config);
+#ifdef USE_TIMING
+  timing.initoutput+=mrun()-t;
+#endif
   failonerror(&config,rc,INIT_OUTPUT_ERR,
               "Initialization of output data failed");
   if(iscoupled(config))
@@ -282,9 +320,15 @@ int main(int argc,char **argv)
     puts("Simulation begins...");
   time(&tstart); /* Start timing */
   /* Starting simulation */
+#ifdef USE_TIMING
+  t=mrun();
+#endif
   year=iterate(output,grid,input,
                config.npft[GRASS]+config.npft[TREE],config.npft[CROP],
                &config);
+#ifdef USE_TIMING
+  timing.iterate+=mrun()-t;
+#endif
   /* Simulation has finished */
   time(&tend); /* Stop timing */
   fcloseoutput(output,&config);
@@ -301,15 +345,6 @@ int main(int argc,char **argv)
            config.total,(int)(tend-tstart),
            (double)(tend-tstart)/config.total/max(year-config.firstyear+
                                                    config.nspinup,1));
-#ifdef USE_TIMING
-    if(iscoupled(config))
-      printf("Time spent in communication to %s model: %.2f sec.\n",
-             config.coupled_model,timing);
-    if(config.restart_filename!=NULL)
-      printf("Time spent in reading restart file: %.2f sec.\n",tread);
-    if(iswriterestart(&config))
-      printf("Time spent in writing restart file: %.2f sec.\n",twrite);
-#endif
   }
 #if defined IMAGE && defined COUPLED
   if(config.sim_id==LPJML_IMAGE)
@@ -324,9 +359,12 @@ int main(int argc,char **argv)
   /* Exit MPI */
   MPI_Finalize();
 #endif
-  time(&tfinal);
+  tfinal=mrun();
   if(isroot(config))
   {
+#ifdef USE_TIMING
+    printtiming(tfinal-tbegin);
+#endif
     printf("Total wall clock time:\t%d sec (",(int)(tfinal-tbegin));
     printtime((int)(tfinal-tbegin));
     puts(").");